Struct signals_rs::Signal

pub struct Signal {
    pub connections: HashMap<String>,
    pub destroyed: bool,
}

signals-rs

signals-rs is a Rust-inspired implementation similar to lua(u) signals/events.

A signal is a global state switch which can be used as a gateway for conditional code execution. Signals can be activated by first “connecting” to them with a callback, then “firing” them to toggle their state.

use signals_rs::Signal;

fn main() {
    let mut some_signal = Signal::new();

    let (_, connection_id) = some_signal.connect(&|| println!("This signal has been fired, continuing..."));
    some_signal.disconnect(connection_id); // This "disconnects" from a signal and removes the registered callback, as it is no longer required.

    some_signal.destroy(); // Signals can be destroyed or dropped too.
}

Signals are especially useful as lightweight events for global state sync-ups.

Fields
connections: HashMap<String>destroyed: bool
Implementations

impl Signal

pub fn new() -> Signal

Signal::new instantiates and returns a new Signal, which can then be made use of.

 use signal_rs::Signal;

 let mut signal = Signal::new();

pub fn connect(&mut self, callback: &'static dyn Fn()) -> (&mut Signal, String)

Signal.connect registers a supplied callback closure/function to be executed on a signal mutation activated by Signal.fire. A signal can be connected to with multiple callbacks.

Every callback returns a tuple with a Signal and Callback ID, which can be stored for future use with Signal.fire & Signal.disconnect.

use signal_rs::Signal;

let mut signal = Signal::new();

let (_, first_callback_id) = signal.connect(&|| println!("received signal fire from callback #1!"));

fn signal_callback() {
    println!("received signal fire from callback #2!");
}

let (_, second_callback_id) = signal.connect(&signal_callback);

println!("#1 -> {}", first_callback_id);
println!("#2 -> {}", second_callback_id);

pub fn disconnect(&mut self, connection_id: Option<String>)

Signal.disconnect disconnects a registered callback from a signal. This prevents execution of a certain callback once it’s been disconnected. An optional connection_id parameter may be provided which can be used to disconnect a specific connection instead of the last accessed connection. Providing a connection_id is highly recommended and minimizes the risk of an unregistered connection from being disconnected.

use signal_rs::Signal;

let mut signal = Signal::new();

let (_, callback_id) = signal.connect(&|| println!("received signal fire from callback"));

signal.disconnect(Some(callback_id));

pub fn fire(&mut self, connection_id: Option<String>)

Signal.fire fires a callback from a connection registered beforehand. An optional connection_id may be provided. In such a case where no connection_id is provided, it will default to the previously accessed connection. It is recommended practice to provide a connection_id as this minimizes the risk of an unregistered connection from being fired.

use signal_rs::Signal;

let mut signal = Signal::new();

let (_, callback_id) = signal.connect(&|| println!("received signal fire from callback"));

signal.fire(Some(callback_id));
signal.disconnect(Some(callback_id));

pub fn destroy(&mut self)

Signal.destroy destroys the signal and all registered callbacks are rendered dysfunctional. It is good practice to destroy a signal once it no longer serves its purpose. Destroying a signal is equivalent to dropping it as the Drop trait has been implemented for Signal.

use signal_rs::Signal;

let mut signal = Signal::new();

signal.destroy()